Add MusicTrigger interface for MusicPredicate and MusicEvent to implement
